continuing work-in-progress
